Refactor pricing and cost calculation methods to use float types; improve error handling in oxygen calculations

+++ casusprogrammeren/Services/Calculation/CostsCalculator.cs
@@ -2,16 +2,16 @@
 
 public class CostsCalculator
 {
-    private static readonly int[] OpeningHoursExtraCharge = [ 5, 4, 3, 2 ];
+    private static readonly float[] OpeningHoursExtraCharge = [ 5, 4, 3, 2 ];
     
-    private static int CalculateExtraCharge(int openingHours )
+    private static float CalculateExtraCharge(int openingHours )
     {
         if (openingHours < 1 || openingHours > 4)
         {
             return 0;
         }
 
-        int total = 0;
+        float total = 0;
         for (int i = 0; i < openingHours; i++)
         {
             total += OpeningHoursExtraCharge[i];
@@ -21,10 +21,10 @@ public class CostsCalculator
         return total;
     }
     
-    public static int CalculateSpectrumRoomCosts(int capacity, int openingHours)
+    public static float CalculateSpectrumRoomCosts(int capacity, int openingHours)
     {
-        int baseCosts = capacity == 60 ? 500 : 300;
-        int extraCharge = CalculateExtraCharge(openingHours);
+        float baseCosts = capacity == 60 ? 500 : 300;
+        float extraCharge = CalculateExtraCharge(openingHours);
     
         if (extraCharge == 0)
         {
@@ -34,10 +34,10 @@ public class CostsCalculator
         return baseCosts + extraCharge;
     }
 
-    public static int CalculatePrismaRoomCosts(int capacity, int openingHours)
+    public static float CalculatePrismaRoomCosts(int capacity, int openingHours)
     {
-        int baseCosts = capacity == 60 ? 475 : 275;
-        int extraCharge = CalculateExtraCharge(openingHours);
+        float baseCosts = capacity == 60 ? 475 : 275;
+        float extraCharge = CalculateExtraCharge(openingHours);
     
         if (extraCharge == 0)
         {
@@ -47,26 +47,26 @@ public class CostsCalculator
         return baseCosts + extraCharge;
     }
 
-    public static int CalculateSpectrumWorkspaceCosts(int openingHours)
+    public static float CalculateSpectrumWorkspaceCosts(int openingHours)
     {
-        int baseCosts = 100;
-        int extraCharge = CalculateExtraCharge(openingHours);
+        float baseCosts = 100;
+        float extraCharge = CalculateExtraCharge(openingHours);
     
         return extraCharge == 0 ? baseCosts : baseCosts + extraCharge;
     }
 
-    public static int CalculatePrismaWorkspaceCosts(int openingHours)
+    public static float CalculatePrismaWorkspaceCosts(int openingHours)
     {
-        int baseCosts = 90;
-        int extraCharge = CalculateExtraCharge(openingHours);
+        float baseCosts = 90;
+        float extraCharge = CalculateExtraCharge(openingHours);
     
         return extraCharge == 0 ? baseCosts : baseCosts + extraCharge;
     }
 
-    public static int CalculatePublicSpaceCosts(int openingHours)
+    public static float CalculatePublicSpaceCosts(int openingHours)
     {
-        int baseCosts = 200;
-        int extraCharge = CalculateExtraCharge(openingHours);
+        float baseCosts = 200;
+        float extraCharge = CalculateExtraCharge(openingHours);
     
         return extraCharge == 0 ? baseCosts : baseCosts + extraCharge;
     }
